requirements.txt
Django==5.2.1
requests==2.32.3
python-dateutil==2.9.0.post0
apify-client==1.7.1
pytz==2024.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0 